+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## unreleased
+
+- Update README.md
+
 ## 0.1.1
 
 - Fix example link in README.md